4)Implementacion/backend/requirements.txt
flask
python-dotenv
flask-restful
flask-jwt-extended
flask_pymongo
certifi
pymongo
flask_mail
flask-login
WTForms
oauthlib
requests
pyOpenSSL



